update constant name
