+++ ts/utils/statusBar.ts
@@ -1,6 +1,4 @@
 import { Platform, StatusBar, StatusBarStyle } from "react-native";
-import { getStatusBarHeight, isIphoneX } from "react-native-iphone-x-helper";
-import customVariables from "../theme/variables";
 
 export const setStatusBarColorAndBackground = (
   style: StatusBarStyle,
@@ -12,11 +10,3 @@ export const setStatusBarColorAndBackground = (
     StatusBar.setBackgroundColor(color, animated);
   }
 };
-
-export const getHeaderHeight =
-  customVariables.appHeaderHeight +
-  (Platform.OS === "ios"
-    ? isIphoneX()
-      ? 18
-      : getStatusBarHeight(true)
-    : customVariables.spacerHeight);
